+++ packages/adapter-interfaces/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @authhero/adapter-interfaces
 
+## 0.64.0
+
+### Minor Changes
+
+- Add form components schemas
+
 ## 0.63.0
 
 ### Minor Changes