+++ packages/drizzle/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @authhero/drizzle
 
+## 0.2.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.64.0
+
 ## 0.2.23
 
 ### Patch Changes